backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
anthropic==0.34.0
python-dotenv==1.0.0
werkzeug==3.0.1
